code/demo/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
tavily-python>=0.3.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
dashscope>=1.14.0
python-docx>=0.8.11
docx2txt>=0.8

